Stricter mixed content check for blob and filesystem URLs
The mixed content spec relies on the concept of potentially trustworthy
URLs which checks trustworthiness of blob and filesystem by checking the
origin of the context in which they were created [1] [2]. Concretely,
"blob:https://" is trustworthy while "blob:http://" is not. Currently
Chromium treats any blob: or filesystem: URLs as trustworthy. This CL
updates the behavior to match the stricter rule from the specification.

[1] https://w3c.github.io/webappsec-mixed-content/
[2] https://w3c.github.io/webappsec-secure-contexts/#is-url-trustworthy
[3] https://groups.google.com/a/chromium.org/g/blink-dev/c/nrpl_ewkmaQ

Bug: 1172988
Change-Id: I6fe6d47d0436d405294194e0fcdd4d48461ad7b5